Started route separation
